[Changed] Make model methods protected on resource query class
The model and modelOrFail methods were never meant to be used
publicly on the resource query class - and were not documented as
a result. To prevent them being used on this class (as it is
confusing as to which model is returned), these methods are now
public on the resource request, and protected on the query request.

See #110